backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
async-timeout==4.0.3
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==3.6.4.0
celery==5.2.7
certifi==2026.1.4
cffi==2.0.0
channels==3.0.5
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==3.0.2
Django==3.2.25
django-celery-beat==2.2.1
django-cors-headers==4.5.0
django-environ==0.13.0
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.15.1
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
multidict==6.7.1
numpy==1.25.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.4.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.9.3
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.4.0
tqdm==4.65.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.2
